+++ Project1_Perpustakaan/src/project1_perpustakaan/DataBuku.java
@@ -451,13 +451,10 @@ public class DataBuku extends javax.swing.JFrame {
     }//GEN-LAST:event_tampilActionPerformed
 
     private void perbaruiActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_perbaruiActionPerformed
-        // Mendapatkan indeks baris yang dipilih
         int selectedRow = tabelDatabaseBuku.getSelectedRow();
 
-        // Memastikan bahwa ada baris yang dipilih sebelum memperbarui
         if (selectedRow != -1) {
             try {
-                // Mendapatkan data dari baris yang dipilih
                 String kodeBukuToUpdate = tabelDatabaseBuku.getValueAt(selectedRow, 0).toString();
                 String judulBukuToUpdate = tabelDatabaseBuku.getValueAt(selectedRow, 1).toString();
                 String namaPengarangToUpdate = tabelDatabaseBuku.getValueAt(selectedRow, 2).toString();
@@ -465,7 +462,6 @@ public class DataBuku extends javax.swing.JFrame {
                 String tahunTerbitToUpdate = tabelDatabaseBuku.getValueAt(selectedRow, 4).toString();
                 String jumlahHalamanToUpdate = tabelDatabaseBuku.getValueAt(selectedRow, 5).toString();
 
-                // Menampilkan data pada JTextField
                 kodeBuku.setText(kodeBukuToUpdate);
                 judulBuku.setText(judulBukuToUpdate);
                 namaPengarang.setText(namaPengarangToUpdate);
@@ -473,67 +469,55 @@ public class DataBuku extends javax.swing.JFrame {
                 tahunTerbit.setText(tahunTerbitToUpdate);
                 jumlahHalaman.setText(jumlahHalamanToUpdate);
 
-                // Menyimpan NIM yang akan diperbarui
-                // Anda bisa menyimpannya sebagai variabel instance jika diperlukan
-                // Contoh: this.nimToUpdate = nimToUpdate;
-
-                // Tambahkan logika lain yang diperlukan untuk proses perbarui
             } catch (Exception ex) {
                 ex.printStackTrace();
             }
         } else {
-            // Menampilkan pesan jika tidak ada baris yang dipilih
             JOptionPane.showMessageDialog(this, "Pilih data yang ingin diperbarui", "Peringatan", JOptionPane.WARNING_MESSAGE);
         }
     }//GEN-LAST:event_perbaruiActionPerformed
 
     private void hapusActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_hapusActionPerformed
-        // Mendapatkan indeks baris yang dipilih
         int selectedRow = tabelDatabaseBuku.getSelectedRow();
 
-        // Memastikan bahwa ada baris yang dipilih sebelum menghapus
         if (selectedRow != -1) {
             try {
-                // Mendapatkan NIM dari baris yang dipilih
                 String kode_bukuToDelete = tabelDatabaseBuku.getValueAt(selectedRow, 0).toString();
 
-                // Menghapus data dari database
-                Connection koneksi = DatabaseConnection.getConnection();
-                String query = "DELETE FROM data_buku WHERE kode_buku = ?";
-                PreparedStatement preparedStatement = koneksi.prepareStatement(query);
-                preparedStatement.setString(1, kode_bukuToDelete);
-                int rowsAffected = preparedStatement.executeUpdate();
+                int option = JOptionPane.showConfirmDialog(this, "Apakah Anda yakin ingin menghapus data ini?", "Konfirmasi", JOptionPane.YES_NO_OPTION);
 
-                if (rowsAffected > 0) {
-                    // Menampilkan pesan sukses dan mengupdate tabel
-                    JOptionPane.showMessageDialog(this, "Data berhasil dihapus dari database", "Sukses", JOptionPane.INFORMATION_MESSAGE);
-                    updateTable();
-                } else {
-                    // Menampilkan pesan kesalahan jika penghapusan gagal
-                    JOptionPane.showMessageDialog(this, "Gagal menghapus data dari database", "Error", JOptionPane.ERROR_MESSAGE);
+                if (option == JOptionPane.YES_OPTION) {
+                    Connection koneksi = DatabaseConnection.getConnection();
+                    String query = "DELETE FROM data_buku WHERE kode_buku = ?";
+                    PreparedStatement preparedStatement = koneksi.prepareStatement(query);
+                    preparedStatement.setString(1, kode_bukuToDelete);
+                    int rowsAffected = preparedStatement.executeUpdate();
+
+                    if (rowsAffected > 0) {
+                        JOptionPane.showMessageDialog(this, "Data berhasil dihapus dari database", "Sukses", JOptionPane.INFORMATION_MESSAGE);
+                        updateTable();
+                    } else {
+                        JOptionPane.showMessageDialog(this, "Gagal menghapus data dari database", "Error", JOptionPane.ERROR_MESSAGE);
+                    }
+
+                    preparedStatement.close();
                 }
-
-                // Menutup resources
-                preparedStatement.close();
             } catch (SQLException | ClassNotFoundException ex) {
                 ex.printStackTrace();
             }
         } else {
-            // Menampilkan pesan jika tidak ada baris yang dipilih
             JOptionPane.showMessageDialog(this, "Pilih data yang ingin dihapus", "Peringatan", JOptionPane.WARNING_MESSAGE);
         }
+
     }//GEN-LAST:event_hapusActionPerformed
 
     private void simpanActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_simpanActionPerformed
         try {
-            // Mendapatkan koneksi dari kelas DatabaseConnection
             Connection koneksi = DatabaseConnection.getConnection();
 
-            // Membuat PreparedStatement untuk menghindari SQL injection
             String query = "UPDATE data_buku SET judul_buku=?, nama_pengarang=?, penerbit=?, tahun_terbit=?, jumlah_halaman=? WHERE kode_buku=?";
             PreparedStatement preparedStatement = koneksi.prepareStatement(query);
 
-            // Mengatur nilai untuk parameter-parameter
             preparedStatement.setString(1, judulBuku.getText());
             preparedStatement.setString(2, namaPengarang.getText());
             preparedStatement.setString(3, penerbit.getText());
@@ -541,17 +525,13 @@ public class DataBuku extends javax.swing.JFrame {
             preparedStatement.setString(5, jumlahHalaman.getText());
             preparedStatement.setString(6, kodeBuku.getText()); 
 
-            // Menjalankan query
             int rowsAffected = preparedStatement.executeUpdate();
 
             if (rowsAffected > 0) {
-                // Menampilkan pesan sukses menggunakan JOptionPane
                 JOptionPane.showMessageDialog(this, "Data berhasil diperbarui di database", "Sukses", JOptionPane.INFORMATION_MESSAGE);
 
-                // Menyegarkan tabel
                 updateTable();
 
-                // Mengosongkan JTextField setelah pembaruan berhasil
                 kodeBuku.setText("");
                 judulBuku.setText("");
                 namaPengarang.setText("");
@@ -560,11 +540,9 @@ public class DataBuku extends javax.swing.JFrame {
                 jumlahHalaman.setText("");
                 cariText.setText("");
             } else {
-                // Menangani kasus jika tidak ada baris yang terpengaruh (pembaruan gagal)
                 JOptionPane.showMessageDialog(this, "Gagal memperbarui data di database", "Error", JOptionPane.ERROR_MESSAGE);
             }
 
-            // Menutup sumber daya
             preparedStatement.close();
         } catch (SQLException ex) {
             ex.printStackTrace();
@@ -575,14 +553,11 @@ public class DataBuku extends javax.swing.JFrame {
 
     private void tambahActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_tambahActionPerformed
         try {
-            // Getting a connection from your DatabaseConnection class
             Connection koneksi = DatabaseKoneksi.DatabaseConnection.getConnection();
 
-            // Creating a PreparedStatement to avoid SQL injection
             String query = "INSERT INTO data_buku (kode_buku, judul_buku, nama_pengarang, penerbit, tahun_terbit, jumlah_halaman) VALUES (?, ?, ?, ?, ?, ?)";
             PreparedStatement preparedStatement = koneksi.prepareStatement(query);
 
-            // Setting values for the parameters
             preparedStatement.setString(1, kodeBuku.getText());
             preparedStatement.setString(2, judulBuku.getText());
             preparedStatement.setString(3, namaPengarang.getText());
@@ -590,11 +565,9 @@ public class DataBuku extends javax.swing.JFrame {
             preparedStatement.setString(5, tahunTerbit.getText());
             preparedStatement.setString(6, jumlahHalaman.getText());
 
-            // Executing the query
             int rowsAffected = preparedStatement.executeUpdate();
 
             if (rowsAffected > 0) {
-                // Display a success message using a JOptionPane
                 JOptionPane.showMessageDialog(this, "Data berhasil dimasukkan ke database", "Sukses", JOptionPane.INFORMATION_MESSAGE);
                 
                 updateTable();
@@ -605,12 +578,9 @@ public class DataBuku extends javax.swing.JFrame {
                 penerbit.setText("");
                 tahunTerbit.setText("");
                 jumlahHalaman.setText("");
-                // You can add additional actions for success here
-
-                // Closing resources
+                
                 preparedStatement.close();
             } else {
-                // Handle the case where no rows were affected (insertion failed)
                 JOptionPane.showMessageDialog(this, "Gagal memasukkan data ke database", "Error", JOptionPane.ERROR_MESSAGE);
             }
         } catch (SQLException ex) {
@@ -627,22 +597,17 @@ public class DataBuku extends javax.swing.JFrame {
     private void cariActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_cariActionPerformed
 
         try {
-            // Mendapatkan koneksi dari kelas DatabaseConnection
             Connection koneksi = DatabaseConnection.getConnection();
 
-            // Membuat PreparedStatement untuk menghindari SQL injection
             String query = "SELECT * FROM data_buku WHERE kode_buku=? OR judul_buku=?";
             PreparedStatement preparedStatement = koneksi.prepareStatement(query);
 
-            // Mengatur nilai untuk parameter
             preparedStatement.setString(1, cariText.getText());
             preparedStatement.setString(2, cariText.getText());
             
-            // Menjalankan query
             ResultSet resultSet = preparedStatement.executeQuery();
 
             if (resultSet.next()) {
-                // Menampilkan data yang ditemukan ke dalam JTextField
                 kodeBuku.setText(resultSet.getString("kode_buku"));
                 judulBuku.setText(resultSet.getString("judul_buku"));
                 namaPengarang.setText(resultSet.getString("nama_pengarang"));
@@ -650,14 +615,11 @@ public class DataBuku extends javax.swing.JFrame {
                 tahunTerbit.setText(resultSet.getString("tahun_terbit"));
                 jumlahHalaman.setText(resultSet.getString("jumlah_halaman"));
 
-                // Menampilkan pesan sukses menggunakan JOptionPane
                 JOptionPane.showMessageDialog(this, "Data ditemukan", "Sukses", JOptionPane.INFORMATION_MESSAGE);
             } else {
-                // Menampilkan pesan jika data tidak ditemukan
                 JOptionPane.showMessageDialog(this, "Data tidak ditemukan", "Peringatan", JOptionPane.WARNING_MESSAGE);
             }
 
-            // Menutup sumber daya
             preparedStatement.close();
         } catch (SQLException ex) {
             ex.printStackTrace();
@@ -686,7 +648,7 @@ public class DataBuku extends javax.swing.JFrame {
             ResultSet resultSet = preparedStatement.executeQuery();
 
             DefaultTableModel model = (DefaultTableModel) tabelDatabaseBuku.getModel();
-            model.setRowCount(0); // Hapus data yang sudah ada di tabel
+            model.setRowCount(0);
 
             while (resultSet.next()) {
                 Object[] row = {
